+++ src/webview/vscode_wrapper.ts
@@ -171,7 +171,7 @@ export class VscodeWrapper {
     switch (e.keyCommand) {
       case 'nextEdge': {controlBar.goToNextTransition(1, []); break;}
       case 'previousEdge': {controlBar.goToNextTransition(-1, []); break}
-      case 'zoomToFit': {this.events.dispatch(ActionType.Zoom, Infinity, 0, 0); break}
+      case 'zoomToFit': {viewport.animateZoomRange(0, viewport.timeStop); break}
       case 'increaseVerticalScale': {this.handleUpdateVerticalScale(e.event, 2); break;}
       case 'decreaseVerticalScale': {this.handleUpdateVerticalScale(e.event, 0.5); break;}
       case 'resetVerticalScale':    {this.handleUpdateVerticalScale(e.event, 0); break;}